Enable parallel build, e.g. with mvnd.
For this, we need to randomize all ports used. In particular, we need to use random ports for all embedded jms brokers (the container brokers are already randomized) and random ports for the http server.

For quarkus-artemis-ra, we do not (yet) have a way to use random ports for the brokers in tests. Therefore, we stay on port 61616 for the broker. Since this is the only broker with a fixed port, collisions with other tests are highly unlikely.

While at it, reduce logging from the embedded artemis server in test execution.